+++ rest_framework/viewsets.py
@@ -44,9 +44,15 @@ class ViewSetMixin(object):
         instantiated view, we need to totally reimplement `.as_view`,
         and slightly modify the view function that is created and returned.
         """
-        # The suffix initkwarg is reserved for identifing the viewset type
+        # The suffix initkwarg is reserved for identifying the viewset type
         # eg. 'List' or 'Instance'.
         cls.suffix = None
+
+        # actions must not be empty
+        if not actions:
+            raise TypeError("The `actions` argument must be provided when "
+                            "calling `.as_view()` on a ViewSet. For example "
+                            "`.as_view({'get': 'list'})`")
 
         # sanitize keyword arguments
         for key in initkwargs:
@@ -92,12 +98,12 @@ class ViewSetMixin(object):
         view.suffix = initkwargs.get('suffix', None)
         return csrf_exempt(view)
 
-    def initialize_request(self, request, *args, **kargs):
+    def initialize_request(self, request, *args, **kwargs):
         """
         Set the `.action` attribute on the view,
         depending on the request method.
         """
-        request = super(ViewSetMixin, self).initialize_request(request, *args, **kargs)
+        request = super(ViewSetMixin, self).initialize_request(request, *args, **kwargs)
         self.action = self.action_map.get(request.method.lower())
         return request