+++ tests/test_model_serializer.py
@@ -26,6 +26,10 @@ class CustomField(models.Field):
     pass
 
 
+class OneFieldModel(models.Model):
+    char_field = models.CharField(max_length=100)
+
+
 class RegularFieldsModel(models.Model):
     """
     A model class for testing regular flat fields.
@@ -66,6 +70,26 @@ class FieldOptionsModel(models.Model):
     default_field = models.IntegerField(default=0)
     descriptive_field = models.IntegerField(help_text='Some help text', verbose_name='A label')
     choices_field = models.CharField(max_length=100, choices=COLOR_CHOICES)
+
+
+class TestModelSerializer(TestCase):
+    def test_create_method(self):
+        class TestSerializer(serializers.ModelSerializer):
+            non_model_field = serializers.CharField()
+
+            class Meta:
+                model = OneFieldModel
+                fields = ('char_field', 'non_model_field')
+
+        serializer = TestSerializer(data={
+            'char_field': 'foo',
+            'non_model_field': 'bar',
+        })
+        serializer.is_valid()
+        with self.assertRaises(TypeError) as excinfo:
+            serializer.save()
+        msginitial = 'Got a `TypeError` when calling `OneFieldModel.objects.create()`.'
+        assert str(excinfo.exception).startswith(msginitial)
 
 
 class TestRegularFieldMappings(TestCase):
@@ -535,3 +559,53 @@ class TestBulkCreate(TestCase):
 
         # Serializer returns correct data.
         assert serializer.data == data
+
+
+class TestMetaClassModel(models.Model):
+    text = models.CharField(max_length=100)
+
+
+class TestSerializerMetaClass(TestCase):
+    def test_meta_class_fields_option(self):
+        class ExampleSerializer(serializers.ModelSerializer):
+            class Meta:
+                model = TestMetaClassModel
+                fields = 'text'
+
+        with self.assertRaises(TypeError) as result:
+            ExampleSerializer().fields
+
+        exception = result.exception
+        assert str(exception).startswith(
+            "The `fields` option must be a list or tuple"
+        )
+
+    def test_meta_class_exclude_option(self):
+        class ExampleSerializer(serializers.ModelSerializer):
+            class Meta:
+                model = TestMetaClassModel
+                exclude = 'text'
+
+        with self.assertRaises(TypeError) as result:
+            ExampleSerializer().fields
+
+        exception = result.exception
+        assert str(exception).startswith(
+            "The `exclude` option must be a list or tuple"
+        )
+
+    def test_meta_class_fields_and_exclude_options(self):
+        class ExampleSerializer(serializers.ModelSerializer):
+            class Meta:
+                model = TestMetaClassModel
+                fields = ('text',)
+                exclude = ('text',)
+
+        with self.assertRaises(AssertionError) as result:
+            ExampleSerializer().fields
+
+        exception = result.exception
+        self.assertEqual(
+            str(exception),
+            "Cannot set both 'fields' and 'exclude'."
+        )